requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv>=1.0.0
Pillow>=10.4.0
requests>=2.31.0
pymongo>=4.6.0
openai>=2.0.0

# ── RAG Pipeline ──────────────────────────────────────────────────────────────
# Embeddings stored in MongoDB Atlas (persistent across deploys).
# OpenAI text-embedding-3-small (1536-dim) via API — no local model needed.
# NumPy for fast cosine similarity search in Python.
numpy>=1.26.0
